+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * [UIBULKED-308](https://issues.folio.org/browse/UIBULKED-308) CSV: Cancel button doesn't stop file uploading
 * [UIBULKED-306](https://issues.folio.org/browse/UIBULKED-306) "Next" button remains disabled for a while after uploading modified file for Bulk edit (CSV approach)
 * [UIBULKED-287](https://issues.folio.org/browse/UIBULKED-287) "Something went wrong" error occurred clicking "Bulk edit" icon from the bulk edit In app form
+* [UIBULKED-310](https://issues.folio.org/browse/UIBULKED-310) Apply to all items checkbox - default values
 * [UIBULKED-289](https://issues.folio.org/browse/UIBULKED-289) Landing page is not cleared switching to "Query" tab after completed Bulk edit
 
 ## [3.0.5](https://github.com/folio-org/ui-bulk-edit/tree/v3.0.5) (2023-03-22)